+++ lib/mangoapps.rb
@@ -13,6 +13,8 @@ require_relative "mangoapps/modules/notifications"
 require_relative "mangoapps/modules/feeds"
 require_relative "mangoapps/modules/posts"
 require_relative "mangoapps/modules/libraries"
+require_relative "mangoapps/modules/trackers"
+require_relative "mangoapps/modules/attachments"
 
 module MangoApps; end
 
@@ -24,3 +26,5 @@ MangoApps::Client.include(MangoApps::Client::Notifications)
 MangoApps::Client.include(MangoApps::Client::Feeds)
 MangoApps::Client.include(MangoApps::Client::Posts)
 MangoApps::Client.include(MangoApps::Client::Libraries)
+MangoApps::Client.include(MangoApps::Client::Trackers)
+MangoApps::Client.include(MangoApps::Client::Attachments)